+++ src/libAtomVM/resources.c
@@ -595,3 +595,28 @@ void *resource_unserialize(struct ResourceType *resource_type, uint64_t serializ
     synclist_unlock(&resource_type->serialized);
     return result;
 }
+
+static void resource_binary_dtor(ErlNifEnv *caller_env, void *obj)
+{
+    struct ResourceBinary *resource_binary = (struct ResourceBinary *) obj;
+    refc_binary_decrement_refcount(resource_binary->managing_resource, caller_env->global);
+}
+
+const ErlNifResourceTypeInit resource_binary_resource_type_init = {
+    .members = 1,
+    .dtor = resource_binary_dtor,
+};
+
+ERL_NIF_TERM enif_make_resource_binary(ErlNifEnv *env, void *obj, const void *data, size_t size)
+{
+    if (UNLIKELY(memory_erl_nif_env_ensure_free(env, TERM_BOXED_REFERENCE_RESOURCE_SIZE) != MEMORY_GC_OK)) {
+        AVM_ABORT();
+    }
+    struct ResourceBinary *resource_binary = enif_alloc_resource(env->global->resource_binary_resource_type, sizeof(struct ResourceBinary));
+    resource_binary->managing_resource = refc_binary_from_data(obj);
+    resource_binary->data = data;
+
+    term result = term_from_resource_binary_pointer(resource_binary, size, &env->heap);
+    refc_binary_decrement_refcount(refc_binary_from_data(resource_binary), env->global);
+    return result;
+}