+++ src/libAtomVM/term.c
@@ -1286,3 +1286,19 @@ term term_from_resource_type_and_serialize_ref(uint64_t resource_type_ptr, uint6
     uint64_t ref_ticks = globalcontext_get_ref_ticks(glb);
     return term_from_ref_ticks(ref_ticks, heap);
 }
+
+term term_from_resource_binary_pointer(struct ResourceBinary *resource, size_t size, Heap *heap)
+{
+    struct RefcBinary *refc = refc_binary_from_data(resource);
+    term *boxed_value = memory_heap_alloc(heap, TERM_BOXED_REFC_BINARY_SIZE);
+    boxed_value[0] = ((TERM_BOXED_REFC_BINARY_SIZE - 1) << 6) | TERM_BOXED_REFC_BINARY;
+    boxed_value[1] = (term) size;
+    boxed_value[2] = (term) RefcBinaryIsResourceManaged;
+    boxed_value[3] = (term) refc;
+    term ret = ((term) boxed_value) | TERM_PRIMARY_BOXED;
+    // Add the resource to the mso list
+    refc->ref_count++;
+    heap->root->mso_list = term_list_init_prepend(boxed_value + REFC_BINARY_CONS_OFFSET, ret, heap->root->mso_list);
+    refc_binary_increment_refcount(resource->managing_resource);
+    return ret;
+}